Added command to remove markers
When adding markers to file, they aren't removed automatically until you
close the IDE. Therefore I added an option to manually remove them. If
selection is a project/folder/package then it will remove all markers
for all sub resources. If selection is a file then it will only remove
the markers for that file.: