Added checks to get output folders and ignore them
This only checks if the project is a C/C++ or Java project and tries to get the output folders. Then when we parse the project it ignores those folders. However the output folder could change when the project is opened in Eclipse and then these changes won't be detected.