Handling mapping on auto-build trigger
However, an important limitation is that if you have a feature that is annotated in a single file, and you remove the annotation we won't know about this, since the 'FeatureContainer' is associated with a feature, but if the feature is deleted, then we won't know which FeatureContainer that is affected and can't update the metrics. Therefore I will refactor so that a FeatureContainer is identified by an IFile instead.