Changed InFileAnnotationParser to use regex matcher
We are now using a regex matcher to find the feature IDs in the
source files. This has performance advantages and makes FeatureDashboard
independent of a specific format for annotations. Previously, the feature
ID had to be surrounded by brackets. Now, group 1 of the regex is
matched to find the feature ID.